Add more documentation and no more --files
Added more documentation for some functions and improved the help text.
Made files mandatory arguments placed without the need of --files.

Also added error logging to be printed at the end of the run so that it
doesn't get lost. Need to figure out a way to print to sterr instead
for clearer output.